CLN: move unique1d to algorithms from nanops (#14919)
TST: consolidate hashtable testing to test_algos.py
